refactor(rule/status): 패키지 구조 리팩토링
